backend/requirements.txt
pandas
numpy
scikit-learn
joblib
requests
fastapi
uvicorn
python-dotenv
